MdeModulePkg/EsrtFmpDxe: Support multiple devices with 0 HardwareInstance
(cherry picked from commit 090642db7ac124c336da72e1954e1fb09e816fb0)

Skip error check if HardwareInstance is 0 as this either means that
FmpVersion < 3 and not supported or,
"A zero means the FMP provider is not able to determine a
unique hardware instance number or a hardware instance number
is not needed." per UEFI specification.

As the FmpInstances are merged and HardwareInstance is not used
remove error check in this case.

Signed-off-by: Jeff Brasen <jbrasen@nvidia.com>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>
(cherry picked from commit 090642db7ac124c336da72e1954e1fb09e816fb0)

Change-Id: Ifbac1085e25679093496ad74c04b765df2fcf2e7