ArmPkg/CpuDxe ARM: Fix page-to-section attribute conversion
(cherry picked from commit 699372d388a796441d938dba4fdb6525b69ea250)

The section-to-page attribute conversion takes the shareability and
execute-never attributes into account, whereas the page-to-section
counterpart does not. The result is that GetMemoryRegionPage () -which
takes a section attribute argument (via *RegionAttributes) that is
ostensibly based on the first page in the range, but differs from the
actual page attributes when converted back- may return with a
RegionLength of zero. This is incorrect, and confuses code that scans a
region by calling GetMemoryRegion () in sequence.

So fix the conversion, and ASSERT () on a non-zero region length.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit 699372d388a796441d938dba4fdb6525b69ea250)

Change-Id: I3eae4949efcdf3d6e05a71c793ea5d0f4fa5b421