ArmPkg/ArmMmuLib ARM: Remove half baked large page support
(cherry picked from commit 82ccaaf8e77fcd461fe06f26b020254226e0f54a)

Large page support on 32-bit ARM is essentially a glorified contiguous
bit where 16 consecutive entries describing a contiguous range with the
same attributes are presented in a way that permits the TLB to cache its
translation with a single entry.

This was never wired up completely, and does not add a lot of value in
EFI, where the page granularity is 4k and we expect to be able to set RO
and XP permissions on individual pages.

Given that large page support complicates the handling of the XN bit at
the page level (which is in a different place depending on whether the
page is small or large), let's just rip it out.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit 82ccaaf8e77fcd461fe06f26b020254226e0f54a)

Change-Id: I5dac4a426aa8b6bfe7c29fa0425b30a6413544bf