ArmPkg/ArmMmuLib ARM: Split off XN page descriptor bit from type field
(cherry picked from commit 3b76284883df833170f741c9fb4037579635b7f2)

With large page support out of the picture, we can treat bits 1 and 0 of
the page descriptor as individual valid and XN bits, instead of treating
XN as a page type. Doing so aligns the handling of the attribute with
the section descriptor layout, as well as the XN handling on AArch64,
and this is beneficial for maintainability.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit 3b76284883df833170f741c9fb4037579635b7f2)

Change-Id: I5d692dd98ef0a4d410a59992cfd5a13e77180597