ShellPkg: Add a condition to print accurate type0 Bios Size in smbiosview.
(cherry picked from commit 78262899d225eb30e5fbe6a88e85a4b1d8c04a61)

As per the SMBIOS spec, in smbios type0 table, if the Bios size is
greater than 16MB, extended bios size is used to update size information
and bios size is set to 0xff. when this data is printed by smbiosview,
both bios size and extended bios size is printed if the smbios version
is beyond 3.1, which is incorrect as Bios size is set to 0xff when
rom size is more than 16MB.

To fix this bug, added a condition to print bios size only when it is
not set to 0xff or if the smbios version is older than 3.1.

Signed-off-by: Thejaswani Putta <tputta@nvidia.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Zhichao Gao <zhichao.gao@intel.com>
Reviewed-by: Zhichao Gao <zhichao.gao@intel.com>
(cherry picked from commit 78262899d225eb30e5fbe6a88e85a4b1d8c04a61)

Change-Id: I02007d02ce7f2d5b0c5f8a71d350a83027663e43