OvmfPkg/VirtioSerialDxe: add driver
(cherry picked from commit 4d1452c59979b9fcc762db692688b5b3b6573106)

Add a driver for the virtio serial device.

The virtio serial device also known as virtio console device because
initially it had only support for a single tty, intended to be used as
console.  Support for multiple streams and named data ports has been
added later on.

The driver supports tty ports only, they are registered as SerialIo
UART in the system.

Named ports are detected and logged, but not exposed as devices.  They
are usually used by guest agents to communicate with the host.  It's not
clear whenever it makes sense for the firmware to run such agents and if
so which efi protocol could be to expose the ports.  So leaving that for
another day.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit 4d1452c59979b9fcc762db692688b5b3b6573106)

Change-Id: Id51b3814ed3316cdd4d9725467a915c56217cb14