UefiCpuPkg: Add DEBUG_CODE for special case when clear RP
(cherry picked from commit 6585ced55858fbab7c2dda8e61b0de99cb891ec9)

In ConvertMemoryPageAttributes() function, when clear RP for a
specific range [BaseAddress, BaseAddress + Length], it means to
set the present bit to 1 and assign default value for other
attributes in page table. The default attributes for the input
specific range are NX disabled and ReadOnly. If there is existing
present range in [BaseAddress, BaseAddress + Length] and the
attributes are not NX disabled or not ReadOnly, then output the
DEBUG message to indicate that the NX and ReadOnly attributes of
the existing present range are modified in the function.

Signed-off-by: Dun Tan <dun.tan@intel.com>
Cc: Eric Dong <eric.dong@intel.com>
Reviewed-by: Ray Ni <ray.ni@intel.com>
Cc: Rahul Kumar <rahul1.kumar@intel.com>
Cc: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit 6585ced55858fbab7c2dda8e61b0de99cb891ec9)

Change-Id: I665b88761aae89a3f8290e20eb7e7208aa4da0c1