OvmfPkg/ResetVector: Fix assembler bit test flag check
(cherry picked from commit e674096accc8e57cd0dd84679905e1222423251e)

Commit 63c50d3ff2854a76432b752af4f2a76f33ff1974 changed the check that is
used to determine if SEV-ES is active. Originally, a CMP instruction with
a supporting JZ instruction was used for the check. It was changed to use
the BT instruction but not JZ instruction. The result of a BT instruction
changes the the carry flag (CF) and not the zero flag (ZF). As a result,
the wrong condition is being checked. Update the JZ to a JNC to properly
detect if SEV-ES is active.

Fixes: 63c50d3ff285 ("OvmfPkg/ResetVector: cache the SEV status MSR...")
Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
(cherry picked from commit e674096accc8e57cd0dd84679905e1222423251e)

Change-Id: Iff355790fa322c7426f574606c0905f03293f98b